requirements_tests.txt
pytest-xdist
numpy==1.22.2
scipy==1.8.0
autograd==1.3
absl-py==0.15.0
dm-tree==0.1.6
dm-haiku==0.0.5
jax==0.2.28
jaxlib==0.1.74
tensorflow==2.7.0
torch==1.7.1
